Hijacked LoadSavedWMStuff.
LoadSavedWMStuff (0x80A5274) was using the vanilla offset for saved WMData. This inserts Stan's replacement for this function, which uses the offset determined by the user.

This should fix the issue where an incorrect chapter title is displayed for save files which were created during a skirmish, which seems to be the only time LoadSavedWMStuff is called.